Added ability to enable or disable logging on a per thread basis and added control commands to turn this on.

+++ includes/iw_cmdline.h
@@ -41,7 +41,7 @@ typedef struct _iw_opt {
     /// The option value.
     union {
         bool           flag;    ///< A flag.
-        long int       num;     ///< A number.
+        long long int  num;     ///< A number.
         char          *str;     ///< A string.
         char           ch;      ///< A character.
     } val;
